Fix Entity

+++ src/Domain/Entities/Order.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using SWD_Laundry_Backend.Domain.Entities.Validation;
 using SWD_Laundry_Backend.Domain.IdentityModel;
 
+
 namespace SWD_Laundry_Backend.Domain.Entities;
 #nullable disable
 
@@ -14,7 +15,7 @@ public class Order : BaseAuditableEntity
 {
     public DateTime OrderDate { get; set; }
     public DateTime ShipDate { get; set; }
-    public decimal Amount { get; set; }
+    public int Amount { get; set; }
 
     /// <summary>
     /// Relationship
